Moving edit headers
