+++ src/scripts/tasks/TasksCreateObject.js
@@ -3,6 +3,21 @@ import TasksFetch from "./TasksFetch"
 import TasksEditForm from "./TasksEditForm"
 
 const TasksCreateObject = {
+  addHeaders () {
+    
+
+    let taskEditHeader = document.createElement("h3");
+    taskEditHeader.textContent = "Click Task to Edit";
+    taskEditHeader.setAttribute("class", "headerH3");
+
+    let taskEditHeader2 = document.createElement("h4");
+    taskEditHeader2.textContent = "Check Box to Mark Complete";
+    taskEditHeader2.setAttribute("class", "headerH4");
+
+    taskFormFragment.appendChild(taskEditHeader);
+    taskFormFragment.appendChild(taskEditHeader2);
+  },
+
    taskBuilder(taskItem) {
 
     let taskArticle = document.createElement("article")
